Add files via upload
This code is attempting to access instance metadata from cloud providers AWS, GCP, and Azure. It does this by making HTTP GET requests to specific URL paths on the target IP address. If the request is successful (i.e. the status code is 200), it prints a message indicating that the metadata is available.

To use this code, you would need to replace the target IP address with the IP address of the instance you want to access metadata from. You can then run the code and it will make the requests and check for the availability of metadata.

Instance metadata is data about an instance that is available to the instance itself. It can include information such as the instance's hostname, IP address, and user data. It is often used to pass configuration information to an instance when it is launched.